Enable avx512 and vnni512 builds for stockfish
Add detection of the AVX512 based features and build stockfish with that.
The vnni256 build is not done as it's the same as vnni512 but with
preferred register width 256. The CPU flags are the same so this cannot
be detected automatically and would have to be selected by user. Given the
results below it does not significantly improve things so it's probably
not worth the effort.

I've tested the build locally with fishnet client, subjectively the
reported knps increased by about 1000 but you don't have to believe
that.

Stockfish benchmark shows a minor NPS improvement for the accelerated
builds compared to BMI2, and if not at least it's not worse.
Sample short tc tournaments and and estimated Elo, calculated by ordo:

tc 1+1, 40 games

    # PLAYER        :  RATING  POINTS  PLAYED   (%)
    1 sf-bmi2       :  2317.6     8.5      16    53
    2 sf-vnni512    :  2317.6     8.5      16    53
    3 sf-vnni256    :  2317.6     8.5      16    53
    4 sf-avx2       :  2282.4     7.5      16    47
    5 sf-avx512     :  2264.7     7.0      16    44

tc 10+0.1, 40 games

    # PLAYER        :  RATING  POINTS  PLAYED   (%)
    1 sf-vnni512    :  2354.0     9.5      16    59
    2 sf-avx2       :  2335.9     9.0      16    56
    3 sf-vnni256    :  2318.0     8.5      16    53
    4 sf-avx512     :  2264.3     7.0      16    44
    5 sf-bmi2       :  2227.8     6.0      16    38

tc 10+1, 40 games

    # PLAYER        :  RATING  POINTS  PLAYED   (%)
    1 sf-avx2       :  2317.6     8.5      16    53
    2 sf-vnni512    :  2300.0     8.0      16    50
    3 sf-avx512     :  2300.0     8.0      16    50
    4 sf-bmi2       :  2300.0     8.0      16    50
    5 sf-vnni256    :  2282.4     7.5      16    47