fix: make `build.rs` idempotent
Cargo build scipts [should not modify any files outside of `OUT_DIR`](https://doc.rust-lang.org/cargo/reference/build-scripts.html#outputs-of-the-build-script).
Both Stockfish and Fairy-Stockfish `Makefile`s generate output files in
their `src` folders, which was making the build non-idempotent. To fix
this, a copy of the repository is created in `OUT_DIR` and the
compilation is performed there.

This is not only a theoretical concern: this was actually failing when
using rust-analyzer with a separate `targetDir`, since the normal and
the analyzer builds can run in parallel, modifying files in the source
folder concurrently and making each other fail.