fix: hint for submodule init on fresh repo
Submodules need to be initialized before building on a fresh repo. There
is an error message indicating this, but #278 introduced a bug where the
warning is not shown anymore.

The root cause is that submodules are (empty) directories even if not
yet initialized so checking if "submodule" is a directory always passes.
Instead it should check "submodule/src", which should only exist if the
submodule has been initialized.